requirements.txt
streamlit==1.3.0
pandas==1.3.4
requests==2.26.0
python-dotenv==0.19.2
plotly==5.3.1
